requirements.txt
absl-py==0.11.0
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
astunparse==1.6.3
backcall==0.2.0
beautifulsoup4==4.9.3
black==20.8b1
cached-property==1.5.2
cachetools==4.1.1
captcha==0.3
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
colorama==0.4.3
decorator==4.4.2
Django==3.1
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.11.1
djoser==2.0.3
factory-boy==3.1.0
Faker==5.0.2
fastapi==0.61.1
fdb==2.0.1
flake8==3.8.3
future==0.18.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
gunicorn==20.0.4
h11==0.10.0
h2==3.2.0
h5py==2.10.0
hpack==3.0.0
httpie==2.3.0
Hypercorn==0.10.2
hyperframe==5.2.0
idna==2.10
importlib-metadata==2.0.0
ipython==7.17.0
ipython-genutils==0.2.0
isort==5.4.2
jedi==0.17.2
Keras==2.4.3
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.4.3
Markdown==3.3.3
mccabe==0.6.1
mypy==0.782
mypy-extensions==0.4.3
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.42
opt-einsum==3.3.0
pandas==1.1.1
parso==0.7.1
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
priority==1.3.0
prompt-toolkit==3.0.6
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pydantic==1.6.1
pydocstyle==5.1.1
pyflakes==2.2.0
Pygments==2.6.1
pylint==2.6.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.11.13
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.6
scipy==1.5.4
six==1.15.0
snowballstemmer==2.0.0
soupsieve==2.1
SQLAlchemy==1.3.19
sqlparse==0.3.1
starlette==0.13.6
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.1
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.11
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
wsproto==0.15.0
zipp==3.4.0
